feat: Add new RPC methods for starting and finishing learning
This commit adds two new RPC methods to the Word service: startNewLearn and finishLearn. These methods allow users to start and finish learning sessions, respectively. The startNewLearn method takes a NewLearnRequest message as input and returns a NewLearnReply message with a list of WordInner objects and an event ID. The finishLearn method takes a FinishLearnRequest message as input and does not return any data.

The WordInner message has been updated to include an extended_blocks field, which was missing in the previous version of the code.

This commit also includes updates to the WordRequest and WordReply messages to use the WordInner message instead of individual fields.

The changes in this commit are aimed at enhancing the functionality of the Word service by providing new methods for learning words.